Move IssueComments in Kantoku out to an individual file
